+++ src/connectedvmware/azext_connectedvmware/custom.py
@@ -160,6 +160,7 @@ def connect_vcenter(
             creds['password'] = pwinput('Please provide vcenter password: ')
             if not creds['password']:
                 print('Parameter is required, please try again')
+                continue
             passwdConfim = pwinput('Please confirm vcenter password: ')
             if creds['password'] != passwdConfim:
                 print('Passwords do not match, please try again')
@@ -981,19 +982,23 @@ def create_vm(
     machine = None
     try:
         machine = machine_client.get(resource_group_name, resource_name)
+        machine_kind = None
         if f"{machine.kind}".lower() != f"{vcenter.kind}".lower():
-            raise InvalidArgumentValueError(
-                "The existing Machine resource is not of the same kind as the vCenter. " +
-                f"Machine kind: '{machine.kind}', vCenter kind: '{vcenter.kind}'"
-            )
+            if machine.kind:
+                raise InvalidArgumentValueError(
+                    "The existing Machine resource is not of the same kind as the vCenter. " +
+                    f"Machine kind: '{machine.kind}', vCenter kind: '{vcenter.kind}'"
+                )
+            machine_kind = vcenter.kind
         if location is not None and machine.location != location:
             raise InvalidArgumentValueError(
                 "The location of the existing Machine cannot be updated. " +
                 "Either specify the existing location or keep the location unspecified. " +
                 f"Existing location: {machine.location}, Provided location: {location}"
             )
-        if tags is not None:
+        if any(x is not None for x in [machine_kind, tags]):
             m = MachineUpdate(
+                kind=machine_kind,
                 tags=tags,
             )
             machine = machine_client.update(resource_group_name, resource_name, m)